Assert if the empty string allocated with `StringStreamAlloc()` and `StringStreamNAlloc()` function starts with a null ('\0') byte
It is a must have to have a null byte ('\0') at the begining of a empty string right
away it is requested by the program using either of the above functions.  This fixes
the problem of erasing the garbage value inside each allocated block for the string as
the first character is ('\0') then the copy operation will always begin from the
beginning of the string.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>