Test the functionality of `JSON_Init*Impl()` functions
These functions come in "Implementation details" but as they are visible to other
modules we've ability to test them so we must assure if they work with basic objects as
their arguments.

This commit does not bring up test suites for `JSON_InitObjectImpl()` function, we will
have it tested in future commit.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>