Test the values of macros of `Map` module
It may happen that while building the library we could mistakenly supply macro
definition over the command line that may change the actual value of these macros
which will result is buggy output.

To fix this we assert the values of these macros using `gtest` library in our automated
tests.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>