Add sanity check in function `MapAllocNBuckets()` to check if the given `hash` and `keycmp` function are not `NULL`
If this sanity check fails then we default the values of function pointer `hash` and
`keycmp` to `Hash` and `KeyCmp` respectively in order to prevent "segmentation fault".

User has the privilage to change the behaviour of the library by providing custom hash
generating functions and key comparing functions on the basis of their requirements but
make sure no one provides `NULL` as de-referencing `NULL` pointer will raise
"segmentation fault".

Also test function `MapAlloc()` once when given `NULL` as arguments and once when given
the built-in functions as arguments.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>