Copy the supplied `key` and `value` to `MapEntry` `key` and `value` to avoid creating a non-constant copy everytime the `key` and `value` needs to be passed to a function
This was a serious design problem in `cjson` and still it is in many other
program files where a non-constant copy is needed to be supplied to a
function before it performs it steps.

This design is really bad design and need to be removed because we need to
create uneccessary non-constant copies of strings before we can supply them
to a function.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>