Add unit tests to test `StringStream` implementation
The unit tests verify that the "StringStream" implementation works correctly,
by checking the behavior of the functions that create, reallocate, and
deallocate "StringStreams". There are tests for different scenarios, such as
creating a new empty "StringStream", creating a "StringStream" from an
existing string, reallocating a "StringStream" to a larger or smaller size,
and deallocating a "StringStream".

The tests also use helper functions, such as
"ComputeStringStreamBufferCapacity", which calculates the required capacity for
a given "StringStream" length.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>