Add unit tests for `StringStreamBegin` and `StringStreamEnd` functions
This commit adds unit tests for the "StringStreamBegin" and "StringStreamEnd"
functions to ensure that they return the correct pointers to the beginning and
end of the data in the "StringStream" object.

These tests will help prevent regressions and ensure that the functions
continue to work as expected even after changes are made to the code.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>