Add `googletest` unit tests for `StringStreamConcat`, `StringStreamRead`, and `StringStreamRetreat` functions
This commit adds new unit tests for the "StringStreamConcat",
"StringStreamRead", and "StringStreamRetreat" functions in the
"sstream/modifiers.c" file. The "StringStreamConcat" function concatenates a
newly formatted string onto the existing string inside the "StringStream"
instance and increments the length attribute. The "StringStreamRead" function
concatenates data of known length onto the existing data instance of
"StringStream" instance. The "StringStreamRetreat" function impedes the
position of the terminate character "\0" by length, or if the length is
greater than the data length of "StringStream" instance, places the terminator
"\0" at beginning.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>