Add googletest unit tests for `StringStreamReadFile` and `StringStreamWriteFile`
"StringStreamReadFile" was tested to ensure that it reads the correct length
of data from the file provided, and if length is 0, it reads the entire file
correctly.

"StringStreamWriteFile" was tested to ensure that it writes data from
"StringStream" to the file provided correctly, with various test cases to
verify the different aspects of the function's behavior. The test suite writes
data to "StringStream" and then writes it to a temporary file using
"StringStreamWriteFile", then reads that file to compare the written data with
the original data.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>