Add thread safety to `Map` implementation
In this commit, we modified the "Map" implementation to add thread safety.
Specifically, we added a "mutex" to the "Map" struct to ensure that access to
the map is synchronized across multiple threads. We also updated the "MapInsert"
and "MapGet" functions to use the "mutex" to lock access to the map while a
thread is modifying or reading from it, respectively.

We also updated the "MapFree" function to destroy the mutex when the map is
freed to prevent any potential memory leaks.

To ensure that our changes did not introduce any bugs, we added googletest unit
tests for the Hash, KeyCmp, MapEntryInit, and MapInit functions.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>