Implement `MapRealloc()` and related tests
This commit introduces a new function `MapRealloc()` which is responsible for
resizing the hash table of the `Map` data structure. The new function ensures
that the data in the existing hash table is properly rehashed and transferred
to the new hash table with a larger capacity.

In addition, this commit also adds a set of new tests to verify the correct
behavior of the `MapRealloc()` function. The tests include checking the correct
allocation and initialization of the new hash table, verifying the correct
transfer of data from the old hash table to the new one, and verifying the
correct resizing of the `Map` data structure.

The `MapRealloc()` function and the related tests have been thoroughly tested
and reviewed, and are considered to be stable and ready for production use.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>