Use `PTHREAD_MUTEX_RECURSIVE` to acquire a mutex for preventing deadlock from happening
Since, "pthread_mutex_lock()" is called twice on the same mutex by the same
thread without an intervening unlock, it causes a deadlock. The first call to
"pthread_mutex_lock()" blocks the thread until it acquires the lock, and the
second call to "pthread_mutex_lock()" also block the thread indefinitely,
waiting for the lock to be released. This results in a situation where the
thread is stuck waiting for itself to release the lock, which is a deadlock.

To avoid this, we use a recursive mutex by initializing the mutex with the
"PTHREAD_MUTEX_RECURSIVE" attribute. This allows a thread to acquire the
mutex multiple times without causing a deadlock, as long as it releases the
mutex the same number of times.

Signed-off-by: Ayush Joshi <ayush854032@gmail.com>